Report usages as bytes, with integer type.

+++ landscape/message_schemas.py
@@ -137,7 +137,7 @@ SWIFT_DEVICE_INFO = Message("swift-device-info", {
     })
 
 SWIFT = Message("swift", {
-    "usages": List(Tuple(Int(), Unicode(), Float(), Float(), Float()))})
+    "usages": List(Tuple(Int(), Unicode(), Int(), Int(), Int()))})
 
 KEYSTONE_TOKEN = Message("keystone-token", {
     "data": Any(Bytes(), Constant(None))
